chore: move prod .js files to /lib dir
