+++ lib/pino-opentelemetry-transport.js
