+++ test/types/pino-opentelemetry-transport.test-d.ts
@@ -2,7 +2,7 @@ import { expectType } from 'tsd'
 import { OnUnknown } from 'pino-abstract-transport'
 import { Transform } from 'stream'
 
-import transport from '../../pino-opentelemetry-transport'
+import transport from '../../lib/pino-opentelemetry-transport'
 
 expectType<Promise<Transform & OnUnknown>>(
   transport({
